Consolidate C regressions into shared runner
