+++ test/doltlite_regression_test_c.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -euo pipefail
 
-echo "=== Checkout Persist Failure Repro ==="
+echo "=== DoltLite Regression Tests ==="
 cc -g -I. -I../src -o doltlite_regression_test_c \
   ../test/doltlite_regression_test_c.c libdoltlite.a -lz -lpthread -lm
-./doltlite_regression_test_c checkout_persist_failure
+./doltlite_regression_test_c all
