+++ test/savepoint_catalog_restore_test.sh
@@ -2,6 +2,6 @@
 set -euo pipefail
 
 echo "=== Savepoint Catalog Restore Repro ==="
-cc -g -I. -I../src -o savepoint_catalog_restore_test \
-  ../test/savepoint_catalog_restore_test.c libdoltlite.a -lz -lpthread -lm
-./savepoint_catalog_restore_test
+cc -g -I. -I../src -o doltlite_regression_test_c \
+  ../test/doltlite_regression_test_c.c libdoltlite.a -lz -lpthread -lm
+./doltlite_regression_test_c savepoint_catalog_restore
